Move WorkspaceWithPartialSemantics to Workspaces.Test.Utilities
