+++ src/Workspaces/CoreTest/SolutionTests/SolutionTestHelpers.cs
@@ -32,18 +32,6 @@ namespace Microsoft.CodeAnalysis.UnitTests
             return workspace;
         }
 
-        public static Workspace CreateWorkspaceWithPartalSemantics(Type[]? additionalParts = null)
-            => new WorkspaceWithPartialSemantics(FeaturesTestCompositions.Features.AddParts(additionalParts).GetHostServices());
-
-        private class WorkspaceWithPartialSemantics : Workspace
-        {
-            public WorkspaceWithPartialSemantics(HostServices hostServices) : base(hostServices, workspaceKind: nameof(WorkspaceWithPartialSemantics))
-            {
-            }
-
-            protected internal override bool PartialSemanticsEnabled => true;
-        }
-
         public static Project AddEmptyProject(Solution solution, string languageName = LanguageNames.CSharp)
         {
             var id = ProjectId.CreateNewId();
